backend/requirements.txt
# Python Dependencies for AIISMS Backend
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
python-decouple>=3.8
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
whitenoise>=6.6.0
supabase>=2.0.0
openai>=1.0.0
groq>=0.4.0
requests>=2.31.0
